Added test for Delete Customer
Added a test to verify that deleting a Customer row deletes the corresponding BillingAddress Row. This is actually enforced by the foreign key constraint in the BillingAddress table, not by the trigger.

With this commit, all test cases that change address data in the Customer table, as opposed to the BillingAddress table, have been written.