Another test for split table refactoring
added a test case for the split table refactoring, specifically, test when adding a Customer  using the new columns and tables.

In doing so, realized that the refactoring needs to add defaults, so added ALTER TABLE to "08 - setup for test of split".  Will need to copy those changes later to the refactoring script.