Completed tests for split-table
Wrote all the tests for the split-table refactoring.  Also changed the refactor script to include the versions of the triggers used in the tests.

Still to do: demos of UPDATE and DELETE. What needs to be done for the contract phase. And is the FOREIGN KEY constraint for shipping address properly handled?