Added steps to complete refactor of split key.
Untested.  Need to run through the final steps, especailly the ALTER TABLEs to ensure that they can be run.

Also left to do. Decide whether to add example of shortening a column, decide to add example involving an INSTEAD OF trigger, and decide use of trigger to understand inserted/deleted contents.